dev: add jwt security

+++ src/services/storage-service.ts
@@ -1,12 +1,20 @@
+import { config } from "@/constants/config";
+import { IJwtPayload } from "@/interfaces/token-payload.interface";
+import { verify, decode, JwtPayload } from "jsonwebtoken";
+
 export class StorageService {
   private static readonly TOKEN_KEY = "token";
+  private static readonly SECRET_KEY = config.secretKey;
 
   static getToken(): string | null {
-    return (
-      localStorage.getItem(StorageService.TOKEN_KEY) ||
-      sessionStorage.getItem(StorageService.TOKEN_KEY) ||
-      null
-    );
+    if (typeof window !== "undefined") {
+      return (
+        localStorage.getItem(StorageService.TOKEN_KEY) ||
+        sessionStorage.getItem(StorageService.TOKEN_KEY) ||
+        null
+      );
+    }
+    return null;
   }
   static setToken(token: string, isRemember = true): void {
     isRemember
@@ -14,13 +22,27 @@ export class StorageService {
       : sessionStorage.setItem(StorageService.TOKEN_KEY, token);
   }
 
-  static clear() {
-    localStorage.clear();
-    sessionStorage.clear();
+  static verifyToken(): IJwtPayload | null {
+    const token = this.getToken();
+    if (!token) {
+      return null;
+    }
+    const decoded = decode(token) as IJwtPayload;
+    if (Date.now() >= decoded.exp * 1000) {
+      return null;
+    }
+    return decoded;
   }
 
-  static clearAndRedirect() {
+  static clear(): void {
+    if (typeof window !== "undefined") {
+      localStorage.clear();
+      sessionStorage.clear();
+    }
+  }
+
+  static clearAndRedirect(route = "/login"): void {
     StorageService.clear();
-    window.location.href = "/login";
+    window.location.href = route;
   }
 }
